Add user auth using Firebase
Github user authentication was built using the Firebase SDK
authentication methods.

A store is assigned an owner on first sign in. The assigned user is the
only one with permissions to edit the Fish inventory.

The login compoonent is used to render the Github login button when a
user is not currently signed in or an error message if the user signs in
to a store he/she is not the owner of.